Comparisons on date time objects
Currently when doing a comparison on a datetime object, the resulting date is a string based on the local machine culture. This makes the library behave differently depending on where it is deployed. A standardized ISO format would make a lot more sense and would make the code platform agnostic.